requirements.txt
asgiref==3.7.2
crispy-bootstrap4==2023.1
Django==5.0.2
django-cleanup==8.1.0
django-cors-headers==4.3.1
django-crispy-forms==2.1
django-filter==23.5
django-phonenumber-field==7.3.0
django-resized==1.0.2
django-rest-framework==0.1.0
djangorestframework==3.14.0
drf-yasg==1.21.7
inflection==0.5.1
packaging==23.2
phonenumbers==8.13.29
pillow==10.2.0
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.4.4
uritemplate==4.1.1